Change pages folder structure
