+++ src/App.tsx
@@ -5,19 +5,19 @@ import { LocalizationProvider } from '@mui/x-date-pickers';
 import { AdapterDateFns } from '@mui/x-date-pickers/AdapterDateFnsV3';
 import Login from './pages/Login';
 import { AmbitionProvider } from './contexts/ambition-context';
-import Ambitions from './pages/Ambitions';
+import Ambitions from './pages/MyWay/Ambitions';
 import { DesiredStateProvider } from './contexts/desired-state-context';
 import { ActionProvider } from './contexts/action-context';
 import { amber, grey, red, teal, orange } from '@mui/material/colors';
-import DesiredStates from './pages/DesiredStates';
-import Actions from './pages/Actions';
+import DesiredStates from './pages/MyWay/DesiredStates';
+import Actions from './pages/MyWay/Actions';
 import { MemoProvider } from './contexts/memo-context';
-import Memos from './pages/Memos';
+import Memos from './pages/Notes/Memos';
 import { TagProvider } from './contexts/tag-context';
 import { ChallengeProvider } from './contexts/challenge-context';
-import Challenges from './pages/Challenges';
+import Challenges from './pages/Notes/Challenges';
 import { ReadingNoteProvider } from './contexts/reading-note-context';
-import ReadingNotes from './pages/ReadingNotes';
+import ReadingNotes from './pages/Notes/ReadingNotes';
 import { ActionTrackProvider } from './contexts/action-track-context';
 import ActionTracks from './pages/ActionTracks';
 import Aggregations from './pages/ActionTracks/Aggregations';
