--- src/pages/Ambitions/Menus/DesiredStateMenu.tsx
@@ -1,123 +1,0 @@
-import { useState } from 'react';
-import { IconButton, Menu, MenuItem, ListItemIcon, ListItemText } from '@mui/material';
-import MenuIcon from '@mui/icons-material/Menu';
-import EditIcon from '@mui/icons-material/Edit';
-import AddIcon from '@mui/icons-material/Add';
-import LinkIcon from '@mui/icons-material/Link';
-import ArchiveIcon from '@mui/icons-material/Archive';
-import ConfirmationDialog from '../../../components/ConfirmationDialog';
-import useAmbitionContext from '../../../hooks/useAmbitionContext';
-import type { DesiredStateWithActions } from '../../../types/desired_state';
-import DesiredStateDialog from '../Dialogs/DesiredStateDialog';
-import ActionDialog from '../Dialogs/ActionDialog';
-import LinkActionsDialog from '../Dialogs/LinkActionsDialog';
-
-interface DesiredStateMenuProps {
-    desiredState: DesiredStateWithActions;
-}
-
-type DialogType = 'Edit' | 'Archive' | 'AddAction' | 'LinkActions';
-
-const DesiredStateMenu = ({ desiredState }: DesiredStateMenuProps) => {
-    const [openedDialog, setOpenedDialog] = useState<DialogType>();
-
-    const [anchorEl, setAnchorEl] = useState<HTMLElement>();
-    const open = Boolean(anchorEl);
-
-    const { archiveDesiredState } = useAmbitionContext();
-
-    const getDialog = () => {
-        switch (openedDialog) {
-            case 'Edit':
-                return <DesiredStateDialog onClose={() => setOpenedDialog(undefined)} desiredState={desiredState} />;
-            case 'Archive':
-                return (
-                    <ConfirmationDialog
-                        onClose={() => setOpenedDialog(undefined)}
-                        handleSubmit={() => {
-                            archiveDesiredState(desiredState.id);
-                            setOpenedDialog(undefined);
-                        }}
-                        title='Archive DesiredState'
-                        message='This DesiredState will be permanently archived. (Linked Ambitions/Actions will not be archived). Would you like to proceed?'
-                        actionName='Archive'
-                    />
-                );
-            case 'AddAction':
-                return <ActionDialog onClose={() => setOpenedDialog(undefined)} desiredState={desiredState} />;
-            case 'LinkActions':
-                return <LinkActionsDialog onClose={() => setOpenedDialog(undefined)} desiredState={desiredState} />;
-        }
-    };
-
-    const handleClick = (event: React.MouseEvent<HTMLButtonElement>) => {
-        setAnchorEl(event.currentTarget);
-    };
-    const handleClose = () => {
-        setAnchorEl(undefined);
-    };
-
-    return (
-        <>
-            <IconButton
-                onClick={handleClick}
-                color='primary'
-                sx={{ position: 'absolute', top: 0, right: 0 }}
-                aria-controls={open ? 'basic-menu' : undefined}
-                aria-haspopup='true'
-                aria-expanded={open ? 'true' : undefined}
-            >
-                <MenuIcon />
-            </IconButton>
-            <Menu anchorEl={anchorEl} open={open} onClose={handleClose} MenuListProps={{ 'aria-labelledby': 'basic-button' }}>
-                <MenuItem
-                    onClick={() => {
-                        setOpenedDialog('Edit');
-                        handleClose();
-                    }}
-                >
-                    <ListItemIcon>
-                        <EditIcon />
-                    </ListItemIcon>
-                    <ListItemText>Edit DesiredState</ListItemText>
-                </MenuItem>
-                <MenuItem
-                    onClick={() => {
-                        setOpenedDialog('Archive');
-                        handleClose();
-                    }}
-                >
-                    <ListItemIcon>
-                        <ArchiveIcon />
-                    </ListItemIcon>
-                    <ListItemText>Archive DesiredState</ListItemText>
-                </MenuItem>
-                <MenuItem
-                    onClick={() => {
-                        setOpenedDialog('AddAction');
-                        handleClose();
-                    }}
-                >
-                    <ListItemIcon>
-                        <AddIcon />
-                    </ListItemIcon>
-                    <ListItemText>Add Action</ListItemText>
-                </MenuItem>
-                <MenuItem
-                    onClick={() => {
-                        setOpenedDialog('LinkActions');
-                        handleClose();
-                    }}
-                >
-                    <ListItemIcon>
-                        <LinkIcon />
-                    </ListItemIcon>
-                    <ListItemText>Link/Unlink Actions</ListItemText>
-                </MenuItem>
-            </Menu>
-            {openedDialog && getDialog()}
-        </>
-    );
-};
-
-export default DesiredStateMenu;
