+++ src/pages/MyWay/Ambitions/Menus/ActionMenu.tsx
@@ -1,0 +1,93 @@
+import { useState } from 'react';
+import { IconButton, Menu, MenuItem, ListItemIcon, ListItemText } from '@mui/material';
+import MenuIcon from '@mui/icons-material/Menu';
+import EditIcon from '@mui/icons-material/Edit';
+import ArchiveIcon from '@mui/icons-material/Archive';
+import ConfirmationDialog from '../../../../components/ConfirmationDialog';
+import type { Action } from '../../../../types/action';
+import useAmbitionContext from '../../../../hooks/useAmbitionContext';
+import ActionDialog from '../Dialogs/ActionDialog';
+
+interface ActionMenuProps {
+    action: Action;
+}
+
+type DialogType = 'Edit' | 'Archive';
+
+const ActionMenu = ({ action }: ActionMenuProps) => {
+    const [openedDialog, setOpenedDialog] = useState<DialogType>();
+
+    const [anchorEl, setAnchorEl] = useState<HTMLElement>();
+    const open = Boolean(anchorEl);
+
+    const { archiveAction } = useAmbitionContext();
+
+    const getDialog = () => {
+        switch (openedDialog) {
+            case 'Edit':
+                return <ActionDialog onClose={() => setOpenedDialog(undefined)} action={action} />;
+            case 'Archive':
+                return (
+                    <ConfirmationDialog
+                        onClose={() => setOpenedDialog(undefined)}
+                        handleSubmit={() => {
+                            archiveAction(action.id);
+                            setOpenedDialog(undefined);
+                        }}
+                        title='Archive Action'
+                        message='This Action will be permanently archived. (Linked Ambitions/DesiredStates will not be archived). Would you like to proceed?'
+                        actionName='Archive'
+                    />
+                );
+        }
+    };
+
+    const handleClick = (event: React.MouseEvent<HTMLButtonElement>) => {
+        setAnchorEl(event.currentTarget);
+    };
+    const handleClose = () => {
+        setAnchorEl(undefined);
+    };
+
+    return (
+        <>
+            <IconButton
+                onClick={handleClick}
+                color='primary'
+                sx={{ position: 'absolute', top: 0, right: 0 }}
+                aria-controls={open ? 'basic-menu' : undefined}
+                aria-haspopup='true'
+                aria-expanded={open ? 'true' : undefined}
+            >
+                <MenuIcon />
+            </IconButton>
+            <Menu anchorEl={anchorEl} open={open} onClose={handleClose} MenuListProps={{ 'aria-labelledby': 'basic-button' }}>
+                <MenuItem
+                    onClick={() => {
+                        setOpenedDialog('Edit');
+                        handleClose();
+                    }}
+                >
+                    <ListItemIcon>
+                        <EditIcon />
+                    </ListItemIcon>
+                    <ListItemText>Edit Action</ListItemText>
+                </MenuItem>
+                <MenuItem
+                    onClick={() => {
+                        setOpenedDialog('Archive');
+                        handleClose();
+                    }}
+                >
+                    <ListItemIcon>
+                        <ArchiveIcon />
+                    </ListItemIcon>
+                    <ListItemText>Archive Action</ListItemText>
+                </MenuItem>
+            </Menu>
+            {openedDialog && getDialog()}
+        </>
+    );
+};
+
+export default ActionMenu;
