+++ src/pages/MyWay/Ambitions/Menus/AmbitionMenu.tsx
@@ -1,0 +1,122 @@
+import { useState } from 'react';
+import { IconButton, Menu, MenuItem, ListItemIcon, ListItemText } from '@mui/material';
+import MenuIcon from '@mui/icons-material/Menu';
+import EditIcon from '@mui/icons-material/Edit';
+import AddIcon from '@mui/icons-material/Add';
+import LinkIcon from '@mui/icons-material/Link';
+import ArchiveIcon from '@mui/icons-material/Archive';
+import ConfirmationDialog from '../../../../components/ConfirmationDialog';
+import type { AmbitionWithLinks } from '../../../../types/ambition';
+import useAmbitionContext from '../../../../hooks/useAmbitionContext';
+import AmbitionDialog from '../Dialogs/AmbitionDialog';
+import DesiredStateDialog from '../Dialogs/DesiredStateDialog';
+import LinkDesiredStatesDialog from '../Dialogs/LinkDesiredStatesDialog';
+
+interface AmbitionMenuProps {
+    ambition: AmbitionWithLinks;
+}
+type DialogType = 'Edit' | 'Archive' | 'AddDesiredState' | 'LinkDesiredStates';
+
+const AmbitionMenu = ({ ambition }: AmbitionMenuProps) => {
+    const [openedDialog, setOpenedDialog] = useState<DialogType>();
+
+    const [anchorEl, setAnchorEl] = useState<HTMLElement>();
+    const open = Boolean(anchorEl);
+
+    const { archiveAmbition } = useAmbitionContext();
+
+    const getDialog = () => {
+        switch (openedDialog) {
+            case 'Edit':
+                return <AmbitionDialog onClose={() => setOpenedDialog(undefined)} ambition={ambition} />;
+            case 'Archive':
+                return (
+                    <ConfirmationDialog
+                        onClose={() => setOpenedDialog(undefined)}
+                        handleSubmit={() => {
+                            archiveAmbition(ambition.id);
+                            setOpenedDialog(undefined);
+                        }}
+                        title='Archive Ambition'
+                        message='This Ambition will be archived. (Linked DesiredStates/Actions will not be archived). Would you like to proceed?'
+                        actionName='Archive'
+                    />
+                );
+            case 'AddDesiredState':
+                return <DesiredStateDialog onClose={() => setOpenedDialog(undefined)} ambition={ambition} />;
+            case 'LinkDesiredStates':
+                return <LinkDesiredStatesDialog onClose={() => setOpenedDialog(undefined)} ambition={ambition} />;
+        }
+    };
+
+    const handleClick = (event: React.MouseEvent<HTMLButtonElement>) => {
+        setAnchorEl(event.currentTarget);
+    };
+    const handleClose = () => {
+        setAnchorEl(undefined);
+    };
+
+    return (
+        <>
+            <IconButton
+                onClick={handleClick}
+                color='primary'
+                sx={{ position: 'absolute', top: 0, right: 0 }}
+                aria-controls={open ? 'basic-menu' : undefined}
+                aria-haspopup='true'
+                aria-expanded={open ? 'true' : undefined}
+            >
+                <MenuIcon />
+            </IconButton>
+            <Menu anchorEl={anchorEl} open={open} onClose={handleClose} MenuListProps={{ 'aria-labelledby': 'basic-button' }}>
+                <MenuItem
+                    onClick={() => {
+                        setOpenedDialog('Edit');
+                        handleClose();
+                    }}
+                >
+                    <ListItemIcon>
+                        <EditIcon />
+                    </ListItemIcon>
+                    <ListItemText>Edit Ambition</ListItemText>
+                </MenuItem>
+                <MenuItem
+                    onClick={() => {
+                        setOpenedDialog('Archive');
+                        handleClose();
+                    }}
+                >
+                    <ListItemIcon>
+                        <ArchiveIcon />
+                    </ListItemIcon>
+                    <ListItemText>Archive Ambition</ListItemText>
+                </MenuItem>
+                <MenuItem
+                    onClick={() => {
+                        setOpenedDialog('AddDesiredState');
+                        handleClose();
+                    }}
+                >
+                    <ListItemIcon>
+                        <AddIcon />
+                    </ListItemIcon>
+                    <ListItemText>Add DesiredState</ListItemText>
+                </MenuItem>
+                <MenuItem
+                    onClick={() => {
+                        setOpenedDialog('LinkDesiredStates');
+                        handleClose();
+                    }}
+                >
+                    <ListItemIcon>
+                        <LinkIcon />
+                    </ListItemIcon>
+                    <ListItemText>Link/Unlink DesiredStates</ListItemText>
+                </MenuItem>
+            </Menu>
+            {openedDialog && getDialog()}
+        </>
+    );
+};
+
+export default AmbitionMenu;
